+++ demos/28-draw-to-svg/index.js
@@ -1,10 +1,5 @@
 /**
- * A script for finding the intersection points of two circles (the 'radical line').
- *
- * Based on the C++ implementation by Robert King
- *    https://stackoverflow.com/questions/3349125/circle-circle-intersection-points
- * and the 'Circles and spheres' article by Paul Bourke.
- *    http://paulbourke.net/geometry/circlesphere/
+ * A script to demonstrate SVG draw and export.
  *
  * @require PlotBoilerplate
  * @require MouseHandler
@@ -13,9 +8,8 @@
  *
  * 
  * @author   Ikaros Kappler
- * @date     2020-09-04
- * @modified 2020-11-13 Added drawing of sector lines.
- * @version  1.0.1
+ * @date     2021-01-04
+ * @version  1.0.0
  **/
 
 
@@ -66,87 +60,35 @@
 	// +---------------------------------------------------------------------------------
 	// | Create a random vertex inside the canvas viewport.
 	// +-------------------------------
-	var randomVertex = function() {
+	/* var randomVertex = function() {
 	    return new Vertex( Math.random()*pb.canvasSize.width*0.5 - pb.canvasSize.width/2*0.5,
 			       Math.random()*pb.canvasSize.height*0.5 - pb.canvasSize.height/2*0.5
 			     );
-	};
+			     }; */
 
-	var circles = [];
-	var circleSegments = [];	
-	
-	for( var i = 0; i < 2; i++ ) {
-	    var center = randomVertex();
-	    var circle = new Circle( center,
-				     i==0
-				     ? Math.abs(randomVertex().x)
-				     : circles[i-1].center.distance(center)*Math.random()*1.2
-				   );
-	    circles[i] = circle;
-	    circleSegments[i] = [ Math.PI*2.0 ];
-	    var radiusPoint = new Vertex( center.clone().addXY(circle.radius*Math.sin(Math.PI/4),circle.radius*Math.cos(Math.PI/4)) );
-	    pb.add( radiusPoint );
-	    pb.add( circle.center );
+	// Use a helper function to build all demo-drawables.
+	var drawables = createDemoDrawables( pb.canvasSize,
+					     'example-image.png',
+					     function() { pb.redraw(); }
+					   );
+	pb.add( drawables );
 
-	    new CircleHelper( circle, radiusPoint, pb );
-	}
 
 	var drawAll = function() {
-	    if( circles.length == 0 ) return;
+	    // TODO: draw everything to SVG
 	    
-	    var radLine = circles[0].circleIntersection( circles[1] );
-	    if( config.alwaysDrawFullCircles || radLine == null ) {
-		pb.draw.circle( circles[0].center, circles[0].radius, 'rgba(34,168,168,0.5)', 1.0 );
-		pb.draw.circle( circles[1].center, circles[1].radius, 'rgba(34,168,168,0.5)', 1.0 );
-	    }
-
-	    if( radLine !== null ) {
-		if( config.drawRadicalLine ) {
-		    pb.draw.line( radLine.a, radLine.b, 'rgba(34,168,168,0.5)', 1.0 );
-		}
-		if( config.drawSectorLines ) {
-		    pb.draw.line( circles[0].center, radLine.a, 'rgba(0,192,192,0.25)', 1.0 );
-		    pb.draw.line( circles[0].center, radLine.b, 'rgba(0,192,192,0.25)', 1.0 );
-		    pb.draw.line( circles[1].center, radLine.a, 'rgba(0,192,192,0.25)', 1.0 );
-		    pb.draw.line( circles[1].center, radLine.b, 'rgba(0,192,192,0.25)', 1.0 );
-		}
-		if( config.drawIntersectionPoints ) {
-		    pb.draw.diamondHandle( radLine.a, 9, 'rgba(0,192,0,1.0)' );
-		    pb.draw.diamondHandle( radLine.b, 9, 'rgba(0,192,0,1.0)' );
-		}
-		if( config.drawCircleSections ) {
-		    drawCircleSection( circles[0], radLine );
-		    drawCircleSection( circles[1], new Line(radLine.b,radLine.a) );
-		}
-	    }
 	};
 
-
-	var drawCircleSection = function( circle, radLine ) {
-	    // Get angle sections in the circles
-	    var lineAa = new Line( circle.center, radLine.a );
-	    var lineAb = new Line( circle.center, radLine.b );
-
-	    var anglea = lineAa.angle();
-	    var angleb = lineAb.angle();
-
-	    var pointa = circle.vertAt(anglea);
-	    var pointb = circle.vertAt(angleb);
-
-	    pb.draw.circleArc( circle.center, circle.radius, angleb, anglea, 'rgba(34,168,168,1.0)', 2.0 );
-	};
-	
-
-	// +---------------------------------------------------------------------------------
-	// | Add a mouse listener to track the mouse position.
-	// +-------------------------------
-	new MouseHandler(pb.canvas,'twocircle-demo')
+	new MouseHandler(pb.canvas,'drawsvg-demo')
 	    .move( function(e) {
 		var relPos = pb.transformMousePosition( e.params.pos.x, e.params.pos.y );
-		var cx = document.getElementById('cx');
+		/* var cx = document.getElementById('cx');
 		var cy = document.getElementById('cy');
 		if( cx ) cx.innerHTML = relPos.x.toFixed(2);
 		if( cy ) cy.innerHTML = relPos.y.toFixed(2);
+		*/
+		stats.mouseX = relPos.x;
+		stats.mouseY = relPos.y;
 	    } );  
 
 
@@ -162,6 +104,10 @@
 	}, GUP );
 	
 
+	var stats = {
+	    mouseX : 0,
+	    mouseY : 0
+	};
 
 	// +---------------------------------------------------------------------------------
 	// | Initialize dat.gui
@@ -173,6 +119,12 @@
 	    gui.add(config, 'drawRadicalLine').onChange( function() { pb.redraw(); } ).name("drawRadicalLine").title("Draw the radical line?");
 	    gui.add(config, 'drawIntersectionPoints').onChange( function() { pb.redraw(); } ).name("drawIntersectionPoints").title("Draw the intersection points?");
 	    gui.add(config, 'drawSectorLines').onChange( function() { pb.redraw(); } ).name("drawSectorLines").title("Draw the sector lines of circle sections?");
+
+	    // Add stats
+	    var uiStats = new UIStats( stats );
+	    stats = uiStats.proxy;
+	    uiStats.add( 'mouseX' );
+	    uiStats.add( 'mouseY' );
 	}
 
 	pb.config.preDraw = drawAll;